+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## [Unreleased]
 
+## [0.2.0] - 2022-03-07
+
+### Add
+
+- Allow aliases of the status tags to be defined with the .define_status_tags
+  method.
+
+### Remove
+
+- No longer define methods with the method_missing hook.
+
 ## [0.1.0] - 2022-03-04
 
 - Initial release